Update SortScore to u16

+++ src/search/sorting.rs
@@ -26,10 +26,10 @@ with this program.  If not, see <http://www.gnu.org/licenses/>.
 use super::Search;
 use crate::{defs::NrOf, movegen::defs::MoveList, movegen::defs::TTMove};
 
-const TTMOVE_SORT_VALUE: u8 = 60;
+const TTMOVE_SORT_VALUE: u16 = 60;
 
 // MVV_VLA[victim][attacker]
-pub const MVV_LVA: [[u8; NrOf::PIECE_TYPES + 1]; NrOf::PIECE_TYPES + 1] = [
+pub const MVV_LVA: [[u16; NrOf::PIECE_TYPES + 1]; NrOf::PIECE_TYPES + 1] = [
     [0, 0, 0, 0, 0, 0, 0],       // victim K, attacker K, Q, R, B, N, P, None
     [50, 51, 52, 53, 54, 55, 0], // victim Q, attacker K, Q, R, B, N, P, None
     [40, 41, 42, 43, 44, 45, 0], // victim R, attacker K, Q, R, B, N, P, None
